Add SPEC0 filtering
This includes two variants of SPEC0 filtering, because there are (at
least?) two interpretations of the text in SPEC0.

One is the literal date of n_months (i.e., 2 years) after release. At
that point, it is no longer supported.

The other is the suggestion in the SPEC of "the next release in a given
quarter is considered as the one removing support for a given item."
This is a little confusing, so my interpretation is the quarter after
the n_months (so you always get at least 2 years).